fix redundant trusted_dealer module
